Unified layer suggestion generation
